Added new general option "--buildFH"
RFC 6145 say: "When the IPv4 sender does not set the DF bit, the translator SHOULD always include an IPv6 Fragment Header to indicate that the sender allows fragmentation.  The translator MAY provide a configuration function that allows the translator not to include the Fragment Header for the non-fragmented IPv6 packets." (page 7)
We didn't provide the configuration function, until now, thats the reason for this commit.