Fixes #125.
The reason why I postponed this so much is because config should not ask session to update the timers, because then they'd be a dependency of each other.
Session needs to stay away from circular dependencies in a panic kind of sense, because it's Jool's most spinlock-crazed module.
The fact that this logic now resides in nl_handler helped clean config from several protocol handling idiosyncrasies, too.