Good chunk of #148.
All the logging is missing is a better timestamp. It's currently just relying on the kernel's time log thingy, which is just number of seconds since the machine was booted, if I recall correctly.
To label issue148 as "Coded" I also need to document the new flags.

To fulfill REQ-12, I divided the solution in two separate flags: `--logging-bib`, which enables logging of only subscriber identifier and external identifier (plus protocol and timestamp), and `--logging-session`, which also logs destination address (and some other garbage).
REQ-12 is `--logging-bib` activated, `--logging-session` deactivated.

I also introduced a versioning system, which ensures the kernel module's version matches the userspace app's.
This is because I've been changing the communication protocol a lot and this is bound to introduce a lot of confusion if people forgets to update one or the other.