Improve success rate of the joolns unit test.
The hang was a panic the kernel was handling surprisingly well.
A missing compilation macro was at fault.

After that, the test was failing somewhat consistently, seemingly
because the kernel was modifying the network namespace
simultaneously behind the test's back. I added a two-second sleep
to reduce collision with other threads.

This is a fix to the latest update to #199.