New NAT64 configuration flag: --handle-rst-during-fin-rcv
If the feature is activated, TCP sessions ended with ungraceful
handshakes are terminated faster.

In the original spec, if a RST is followed by a FIN (to kill the
connection quickly), the session hangs in one of the`FIN RCV` states for
as long as the maximum session lifetime. As a result, abandoned sessions
tend to strangle pool4 usage.

The alternate mode hangs the session in the `FIN RCV` state for as long
as the established connection *idle* timeout, which is roughly similar
to what happens during graceful FIN handshakes.