Fix a few indentation-related compilation warnings
It appears that whoever wrote the JSON parser had very misled
formatting conventions. I cannot picture anyone indenting code like
this having good intentions:

	if (<condition>) <action-if-true>; <action-always>

Not sure what to make of this.

In other news, I finally managed to properly test #245. Though the
EAMT itself might err (as this was outside of the scope of 245),
the code that reviews it should detect certain types of improper
RCU API usage.