BIBless NAT64 prototype
Information on this thing can be found in the mailing list.
Basically, it's a NAT64 that deals with sessions but not BIBs.
This makes it possible to serve a massive number of IPv6 clients
with minimal IPv4 addresses, at the cost of being completely unable
to open connections from IPv4.