Improve mask selection algorithm's performance
This solution is not ideal, but should get the job done. (Again,
assuming that my theory is correct.)

Here's some context:
https://mail-lists.nic.mx/pipermail/jool-list/2017-September/000143.html

I said in the mail that, ideally, the NAT64 would say "if I expect
on average to iterate more than 100 times, pretend that pool4 is
exhausted". Turns out that even something as simple as being able
to tell whether pool4 has been completely exhausted or not is very
difficult. So instead, I'm going with a more reactive solution.

The solution is, "if I have already done 100 iterations, pretend
that pool4 is exhausted".

The number (100 above) can be tweaked. The argument is
--max-mask-iterations. Defaults to 64.

Not as elegant, but hey. Much better than nearing 16 * 64k
iterations when nearing exhaustion.

This strategy capitalizes of the fact that algorithm 3 of RFC 6056
scatters mask allocations randomly. This means that, on average,
the allocatable masks should be more or less uniformly distributed
through pool4.

For example, on a pool with 1000 transport addresses, if the pool
has been 99% exhausted (meaning, there are 10 addresses left),
we expect to iterate about 100 times to find an available address.
Again, because they are somewhat uniformly distributed. Iterating
more is a bad idea, because the probabilities of finding an
available mask are slim. (Especially because finding a full pool4
is way more likely than finding an almost full one.)

This will lead to slight pool4 underutilization, but will
completely trump the performance spike. No matter how exhausted
pool4 is, the code will only iterate --max-mask-iterations times at
worst.

If --max-mask-iterations is zero, there will be no limit, and
therefore Jool will behave as it used to.