Improve the optimization solution (#251)
The iteration limit introduced a few commits back is now configured
per pool4 table, rather than globally. This is because the pool4
tables are not all guaranteed to have the same size.

Also, tweak the iteration limit's default value so it depends on
the table size. Same reason as above.

The only problem is that I feel it's somewhat more complicated now.
Well, at least users don't need to worry about this until they want
to micro-optimize.

There's also a refactor of the userspace app code that got a little
out of control. The code was really messy. It still kind of is, but
far less so.

The way the userspace app shows the default value is confusing;
will fix tomorrow.