Add tool for computing optimal max-iterations for pool4
I'm still not entirely sure if I should commit to the default value
range proposed. I should base it upon real data, but it's proven to
be really slow to compute. So I made this tool for simulating pool4
utilization quickly.

Next commit will be the final work on issue #251. (except for
documentation.)