Divide timer into session and global timer
Due to a possible bottleneck scenario, the session timer needed to be
sepparated from the other timers in order to have control on the
execution time of such timer. The scenario is this: while the sessions
are being removed, the package translation could be delayed by the
execution of the sessions removal task. This is a massive trafic
scenario that could happen when there's a huge amount of expired
sessions to remove.