Modernize the collection of statistics
Atempting to adapt the Filtering module to the new framework, I
realized that there is a rather large amount of paperwork that
needs to be done whenever Jool stops translating a packet:

- Log message
- Statistics update
- Potential ICMP error
- Packet deallocation
- Return value

The old model to do this is no longer as pleasant since we're now
free from Netfilter's idiosyncrasies. I should probably nail it
more elegantly before doing more high-level surgery.

This commit is the beginning of a refactor of the statistics
system. The previous means to collect stats was overly generic, so
I feel that not much information could be extracted from the
collected data. Also, I would like users to be able to review the
information through simple userspace app commands. The fact that
Jool used to update kernel counters wasn't even documented. Stats
should also become a faster troubleshooting option than enabling
debug logging.