Miscellaneous needed work
1. Refactor the packet init routines

The v4 code wasn't a mirror of the v6 code enough, so the debugging
was more confusing that it needed to be.

2. Adding ICMP errors back to the packet failure paperwork

Which involves a RFC 6146 compliance review.

3. Add memory leak validation for error code paths

All error code paths need to end in a deep breakdown() call,
otherwise the post-failure paperwork is not done. But technically
nothing can properly guarantee that this is done for all code
paths.

The deallocation of the packet is by far the most critical part
of the paperwork, and thankfully it's very easy to detect when
we've forgotten it.

So at least the code will validate that.