Fix more broken tests
There are still some tests with warnings due to the "xlator" calls
when the session timers are removed; I've taken this as a general
warning: probably the BIB shouldn't be getting the "xlator".

I'm analizing what to do, since this is probably not the best solution
(to get the "xlator" from the BIB scope).