Change project license to GPLv2
Whether we will prepend the license header to every file is still
TBD, though the release of 3.5.6 has been green-lighted. This will
do for now.

I'm not sure if I tweaked the DKMS packaging copyrights right
because I can't find the documentation that was originally used to
generate them.
I understand that it's undergoing heavy refactors so this shouldn't
be a problem.

Fixes #258.