Refactor the living hell out of the userspace app
There was a lot of mismatch between the grammar expected from the
app arguments and the means through which argp expects to handle
them, which led to a lot of clumsy, patchy code.

Actually, there's a bit of a lot going on in this commit:

- There's the drive to build a jool lib so the translator can be
  easily ported to other frameworks. I separated the argp parsing
  from the netlink request building in preparation for this
  purpose.
- I'm building a module for centralized management of global config
  values. This is because the relevant code was very poorly
  designed, which led to too much paperwork needed whenever a new
  global field was needed.

The code is largely untested and very likely unstable.

--file and --joold have not been adapted:

- I left --file alone because getting the JSON parser to cooperate
  with the global central module is getting a little tricky. Will
  patch --file back in the near future.
- --joold is a royal pain in the ass an I'm thinking about
  releasing the first iteration of Jool 4 without it.

Largely done coding #259.