More issue #247 review
The code wasn't catching NULL as a potential result from
skb_pull().

In all fairness, nobody does, really. Standard skb operations
usually crash the kernel on failure; I don't really understand why
this specific error condition yields NULL instead of BUG(). And no
other kernel citizens seem to care about it.

Also, this is obviously not the cause of #247; it would have to
crash later if this were the case.