Refactor the userspace clients
Instance naming and mandatory NAT64 pool6 made bare argp much too
convoluted for my patience. Jool 4 is likely not going to see the
light of day (because Device Driver Jool is just going to be
another iteration of Jool 3 now), and since we're changing minor
version number in this release, now is the time.

- Remove double dashes in many command line arguments; issue #259.
- Centralize global variable management. Removes a lot of paperwork
  every time we need to add/change/remove some global value.
- Separate the original massive argp parsing into one argp parsing
  per [mode, operation] tuple. Lasagna code FTW.

There are still bugs lurking around.