Remove usage of the INETFRAGS_HASHSZ macro
So apparently, Ubuntu (or rather, Debian?) is now pulling off the RHEL
crap of claiming to be using one kernel, while actually using API from
another.

In all fairness though, INETFRAGS_HASHSZ was probably always meant
to be private. (It's hard to tell really.)

As a side effect, Jool 3.5 now compiles (and runs fine) in kernels 4.17
and 4.18.

Fixes #283.