Patch -Wall -pedantic warnings
I hadn't noticed that these flags were absent from the userspace tools
makefiles. I didn't see any errors, though.

However, the warnings made me realize that the userspace logging
utilities needed a refactor, and this makes up the bulk of the diff.
The jist of it is that the client and joold logging are conceptually
different now, and so a common API might give people the wrong idea.
The APIs are distinct now.

This is the logging status as of now:

- kernel modules: Should log via the log_* API.
  (Which are just wrappers on top of the kernel's official logging
  functions, adding newlines.)
- Userspace clients and iptables shared objects: No logging; these
  should write into the output streams. (stdout and stderr)
  The userspace clients can use argp/log.h, which contains some printing
  functions that automate the stupid newlines.
- joold: syslog
- libjoolutil and libjoolnl: No logging. Any error messages should be
  returned to calling code so the upper layer (clients or joold) can
  print them according to their needs. See util/result.h.