Remove `file remove`, and replacement `-f`
This (which was used to remove the instance whose name was contained in
"<file>"):

        jool file remove <file>

No longer exists. It is now written like this:

        jool -f <file> instance remove

I want this because instance removal is not the only operation that
might need to extract the instance name from a file. Being able to
specify instance name by file during the initial batch of argp is more
versatile; it allows for things like this:

	jool -f <file> global display

This is useful for init scripts, since we don't want to hardcode the
instance name anywhere there.

As a side effect, this commit also improves the parsing of the first
batch of arguments. It used to be handled manually, which made it
somewhat nonstandard. It is now handled by getopt_long(3), and the
overall command line syntax looks like this:

	(jool | jool_siit) <argp1> <mode> <operation> <argp2>