Implement translation of ICMP Extensions (IE)
Where "translation" of an IE simply means "treated as an opaque
but indivisible bit string meant to be either copied to a rightful
place or chopped off completely." The ICMP header length is also
updated accordingly.

Previously, Jool used be unbeknownst of IEs, which means it
treated them like the rest of the ICMP error payload (trimming it
arbitrarily sometimes), and zeroizing the ICMP header length.

I keep forgetting that these requirements exist because they are
positioned akwardly in RFC 7915. This is work that should have
been implemented as part of #221.

This code is untested still.