Purge `force-slow-path-46`
Turns out it doesn't work as intended. allocate_slow() assumes the
resulting packets will have fragment headers, which is not always
true when the user wants to force slow path.

But removing the assumption would require further modifications and
testing. I don't think the help we might get from the flag is worth
the effort. Users can downgrade to 4.0.9 if things go south.