MAP-T development checkpoint 2
(Most of these notes apply to both the previous commit and this one.)

The address translation algorithm is mostly done. Off the top of my
head, the code is missing

- A proper table implementation of the FMR. (There's only one FMR slot
  right now.)
- The Netlink exchange protocol for the MAP-T fields.
- I haven't really given any thought as to how I'm going to connect this
  to NAPT.

I was originally planning to reuse the jool_siit module for MAP-T, but
it hadn't dawned on me that the address translation algorithm needs L4
data. MAP-T requires defragmentation, so it will have to be a separate
binary.

Branch is still unstable.