MAP-T dev checkpoint 7: Reflect latest RFC interpretations
- Remove configuration option `k`

I've been going back and forth as to whether `k` equals `q` or not.
Currently, my stance is `k = q`. This is born from the observation that
the Port-Restricted Port field is nothing but an conceptual helper that
only makes sense when linked to the MAP address format, and the PSID is
their bridge. If `k != q`, the Port-Restricted Port field exists in a
vacuum.

Therefore, `k = q`.

`q = o - p`, and both `o` and `p` are configuration options, so `k`
itself is redundant configuration.

- Revert Extended Port-Restricted Port fields

`a + k + m` cannot be `> 16`, because the Port-Restricted Port field
only exists when `q > 0` (`o + r > 32`). There's no need for PSIDs when
each CE is assigned more than one IPv4 address.

- Add `o + r < 32` unit tests

Patched two bugs in the MAP-T code as a result.