MAP-T dev checkpoint 8: Add MAP-T intro documentation
This is the beginning of what I expect will be an ardous documentation
refactor.

I've meant to bring 464XLAT and SIIT-DC to the introduction, since they
are basically essential, yet their dedicated pages haven't aged well.
Now I have to explain MAP-T, which is a basic and compound translation
mechanism at the same time, so it's the perfect time to do so.

Also, I think old explanations will have to be expanded for greater
detail, and to ease the introduction of new complications brought to the
table by the new translation mechanism.