Hairpinning: Reinforce state object initialization on SIIT
SIIT Jool was not zeroing the hairpinning translation state object on
initialization.

I don't actually know if this yielded misbehavior, but it was
inconsistent with the way the normal packet pipeline, as well as the
hairpinning NAT64 pipeline initialized their state objects.

So I don't know if this is a bugfix, but the code is certainly more
bug-resistant now.