Define the netlink header magic string in the preprocessor.
The magic string was introduced to the netlink header struct in 58bf14e as part
of the fix for NICMx/Jool#340, initially as a hard-coded byte sequence. This
commit moves the magic string and its length into a preprocessor definition,
and reads and writes this field using memcmp() and memmove() -- if the string
ever needs to be changed in the future, then the change will be automatically
picked up by all the code which reads or writes this header field.