Netlink: Simplify kernel module's Netlink core API
Looking for #352 by code analysis, I was hampered by the cluttered
Netlink API. Managing translator instances and Netlink responses
separately lead to overcoding for no practical benefit, and managing
error messages in yet another separate module was just confusing for no
reason. (And dangerous, since it was storing the errors in a global
variable. This didn't cause any errors that I'm aware of, though.)

This centralizes all of that into a request handling state structure,
jnl_state, which lives in the heap, and which mirrors the translation
state structure. (xlation, which would benefit from a rename now.) This
leads to a better substance-to-boilerplate ratio in the Netlink
handlers.

Unfortunately, this is just developer vanity. While I certainly prefer
the new API, I didn't find #352 in the end, or any other errors for that
matter.