7915: Fix checksums on Slow Path
If the theory is correct, this is what's causing #366 and #375.

If input was type CHECKSUM_PARTIAL, Slow Path yielded CHECKSUM_NONE and
the checksum itself was a weird mush between a negative IPv4
pseudoheader and a positive IPv6 pseudoheader. Therefore, the receiver
was probably dropping the packets.

If input is CHECKSUM_PARTIAL now, Slow Path computes the outgoing
checksum from scratch. It's slow but unavoidable, because Slow Path
generates separate fragments, and tx-checksumming doesn't know how to
checksum them. I also don't want to use frag_list because there's little
to no API, and the code is complicated enough as it is.

------------------------------------------------------------------------

I was able to reuse the code to also fix a long-standing TODO: If
IPv4/UDP packet contains zero checksum, instead of computing the
IPv6/UDP checksum manually, Jool now offloads it to the NIC. Gains a
slight performance boost.