F1x #378
Finally.

It was just uninitialized memory showing up as a nonzero initial
counter. Presumably, it only happens when packaged because of the
optimization flags Debian defaults.

And, because the bug normally only appears when packaged, this bug
should not be closed until the release is out.

READY FOR RELEASE!!111!!1!!!11!!!!1oneone

Ugh