Modernize the iptables shared object exports
Fixes Debian bug 1029268:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1029268

man 3 dlopen:

> Use of _init and _fini is now deprecated in favor of the
> aforementioned constructors and destructors, which among other
> advantages, permit multiple initialization and finalization functions
> to be defined.

Replace _init() with __attribute__((constructor)). Vincent Bernat
already confirmed this works.

Also returns the static keyword, which was removed during #337. I'm
conjecturing that patch likely was a misled accident, and the current
one should be the proper fix for both bugs. This, I'm not completely
sure will work, but we'll see.