Put joold->queue sessions to use
joold->queue is a listing of joold sessions whose fetch needs to be
postponed because of the Netlink channel's limitations. Quite
surprisingly, the code was not actually ever fetching them, which is
why they were queuing indefinitely.

It suspect this has gone unnoticed because, depending on the
environment, Joold seemingly needs lots of traffic to queue. I myself
had to disable ss-flush-asap to be able to replicate the bug.

Likely f1xes #410, but I noticed a couple extra issues that need to be
addressed before a new release. It seems joold has gotten dusty again.
It makes me wonder if people is using it.

In particular, I had to disable --advertise to prevent it from
synchronizing uninitialized memory. Will need to look into this in the
following days.