Fix joold advertise, abandoned attempt
For #410.

I more or less finished the implementation (albeit not the testing),
but the unit tests revealed that one of my assumptions regarding Generic
Netlink is incorrect.

nla_put() cannot be trusted to respect the allocated packet size,
because alloc_skb() can reserve more tail area than it's requested.
This means ss-max-payload has to be enforced manually.

And this by itself wouldn't be enough to justify another rewrite, but
after thinking about it, I realized a more natural implementation would
also reduce the range of the spinlock... which is currently not the
best. Lots of performance to be gained from switching.

So checkpointing. I might have to return to this implementation if the
new one turns out unviable.