Fix joold advertise
Had to rewrite kernelside joold again. New, better design. Implements
joold advertise (because it somehow used to be a no-op), while keeping
busy looping and packet allocations outside of the spinlock.

Deprecates ss-max-payload in favor of ss-max-sessions-per-packet,
partly because the latter is more intuitive (hopefully), and partly
because the former was trickier with the new implementation.

Also, please note that the ss-capacity warning changed:

> joold: Too many sessions deferred! I need to drop some; sorry.

Also tweaked the documentation a little. For some reason, it was
parroting that the channel between joolds is TCP, when it's supposed to
be UDP. Also patched some broken links.

Fixes #410.