Reduce severity of joold's desync message
Weird. According to the output, Jool upgraded the error to kernel WARN()
because joold employed log_err() during a soft irq, which is supposed to
break log_err()'s contract. But the relevant code is a userspace request
handler. I'm going to have to research this more.

Also, the error message was too long, and log_err() was truncating it
to 256 characters. log_warn_once() fixes that for now.

Also raise Jool's version, so suffieldacademy can more easily tell what
they're running.