Drop the remaining usages of the skb control buffer
Apparently, IPCB(skb)->frag_max_size is not populated when the fragment
is a ping, at least in one environment. This is causing packet drops.
Therefore, I decided to always autocompute frag_max_size manually
instead.

(I guess, even if I try to check frag_max_size != 0, there's always the
possibility that I'm actually reading garbage because the control buffer
is a different struct.)

This was the last remaining use of the control buffer. This thing has
always turned out to be a shot in the foot...

So, never use the control buffer for anything again.

Thanks to Laura Zelenková for her bug report.